Change "LD A,(n)" and "LD (n),A" and add LDX
Up until now we had the mnemonic LDH for instructions 0xe0 and 0xf0 to
tell them apart from 0xea and 0xfa. In the 1996 Sharp Microcomputer Data
Book they use LDX for 0xea and 0xfa; and LD for 0xe0 and 0xf0.

This commit changes that to comply with the Sharp Data Book. LDH is
still valid as a mnemonic, but LD can't be used for 0xea and 0xfa any
more, which will brake assembly files that use those instructions.